User roles + CSS styling

+++ blog/views.py
@@ -1,4 +1,4 @@
-from django.shortcuts import render,redirect
+from django.shortcuts import render,redirect,HttpResponse
 from .models import Post
 from .forms import AddPostForm,UserRegistrationForm
 from django.contrib.auth.decorators import login_required
@@ -9,7 +9,7 @@ from django.contrib.auth.decorators import login_required
 
 @login_required
 def display_post(request):
-    posts = Post.objects.all()
+    posts = Post.objects.all().order_by('-created_at')
     return render(request,"posts.html",{"posts":posts})
 
 @login_required
@@ -17,7 +17,9 @@ def add_post(request):
     if request.method == "POST":
         form = AddPostForm(request.POST)
         if form.is_valid():
-            form.save()
+            post = form.save(commit=False)
+            post.author = request.user
+            post.save()
             return redirect('display_post')
     else:
         form = AddPostForm()
@@ -27,6 +29,8 @@ def add_post(request):
 @login_required
 def edit_post(request, id):
     post = Post.objects.get(id=id)
+    if post.author != request.user:
+        return HttpResponse("You are not allowed to edit this post.")
 
     if request.method == "POST":
         form = AddPostForm(request.POST, instance=post)
@@ -41,6 +45,8 @@ def edit_post(request, id):
 @login_required
 def delete_post(request, id):
     post = Post.objects.get(pk=id)
+    if post.author != request.user:
+        return HttpResponse("You are not allowed to delete this post.")
     if request.method == 'POST':
         post.delete()
         return redirect('display_post')
